+++ packages/blade/src/components/Dropdown/docs/DropdownWithButton.stories.tsx
@@ -2,17 +2,21 @@ import React from 'react';
 import { DropdownButton } from '../DropdownButton';
 import { Dropdown, DropdownLink, DropdownOverlay } from '..';
 import { DropdownFooter, DropdownHeader } from '../DropdownHeaderFooter';
+import { DropdownIconButton } from '../DropdownIconButton';
 import {
   WithControlledMenuStory,
   WithControlledMultiSelectMenuStory,
   WithLinkStory,
   WithAutoPositioningMenuStory,
   WithSimpleMenuStory,
+  WithIconButtonStory,
+  WithTooltipStory,
 } from './stories';
 import { Sandbox } from '~utils/storybook/Sandbox';
 import { Box } from '~components/Box';
 import { ActionList, ActionListItem, ActionListItemIcon } from '~components/ActionList';
 import {
+  BoxIcon,
   CheckIcon,
   ChevronDownIcon,
   ChevronUpIcon,
@@ -26,6 +30,7 @@ import { Button } from '~components/Button';
 import { Badge } from '~components/Badge';
 import { Amount } from '~components/Amount';
 import { AutoComplete } from '~components/Input/DropdownInputTriggers';
+import { Tooltip, TooltipInteractiveWrapper } from '~components/Tooltip';
 
 const DropdownStoryMeta = {
   title: 'Components/Dropdown/With Button and Link',
@@ -66,6 +71,14 @@ export const WithLink = (): React.ReactElement => {
   );
 };
 
+export const WithIconButton = (): React.ReactElement => {
+  return (
+    <Sandbox padding="spacing.0" editorHeight="100vh">
+      {WithIconButtonStory}
+    </Sandbox>
+  );
+};
+
 export const WithAutoPositioning = (): React.ReactElement => {
   return (
     <Sandbox padding="spacing.0" editorHeight="100vh">
@@ -86,6 +99,14 @@ export const WithControlledMultiSelect = (): React.ReactElement => {
   return (
     <Sandbox padding="spacing.0" editorHeight="100vh">
       {WithControlledMultiSelectMenuStory}
+    </Sandbox>
+  );
+};
+
+export const WithTooltip = (): React.ReactElement => {
+  return (
+    <Sandbox padding="spacing.0" editorHeight="100vh">
+      {WithTooltipStory}
     </Sandbox>
   );
 };
@@ -298,6 +319,55 @@ InternalLinkDropdown.parameters = {
   chromatic: {
     disableSnapshot: false,
   },
+};
+
+export const InternalIconButtonDropdown = (): React.ReactElement => {
+  const [status, setStatus] = React.useState<string | undefined>('latest-added');
+  const [isDropdownOpen, setIsDropdownOpen] = React.useState(false);
+
+  return (
+    <Box padding="spacing.10">
+      <Tooltip content="Check Status">
+        <TooltipInteractiveWrapper>
+          <Dropdown onOpenChange={setIsDropdownOpen} isOpen={isDropdownOpen}>
+            <DropdownIconButton icon={BoxIcon} accessibilityLabel="Status Dropdown" />
+            <DropdownOverlay>
+              <ActionList>
+                <ActionListItem
+                  onClick={({ name, value }) => {
+                    console.log({ name, value });
+                    setStatus(name);
+                  }}
+                  isSelected={status === 'latest-added'}
+                  title="Latest Added"
+                  value="latest-added"
+                />
+                <ActionListItem
+                  onClick={({ name, value }) => {
+                    console.log({ name, value });
+                    setStatus(name);
+                  }}
+                  isSelected={status === 'latest-invoice'}
+                  title="Latest Invoice"
+                  value="latest-invoice"
+                />
+
+                <ActionListItem
+                  onClick={({ name, value }) => {
+                    console.log({ name, value });
+                    setStatus(name);
+                  }}
+                  isSelected={status === 'oldest-due-date'}
+                  title="Oldest Due Date"
+                  value="oldest-due-date"
+                />
+              </ActionList>
+            </DropdownOverlay>
+          </Dropdown>
+        </TooltipInteractiveWrapper>
+      </Tooltip>
+    </Box>
+  );
 };
 
 export default DropdownStoryMeta;