+++ packages/blade/src/components/Elevate/index.ts
@@ -1,0 +1,2 @@
+export { Elevate } from './Elevate';
+export type { ElevateProps } from './types';