+++ packages/blade/src/components/Icons/TranslateIcon/TranslateIcon.native.test.tsx
@@ -1,0 +1,11 @@
+import TranslateIcon from '.';
+import renderWithTheme from '~utils/testing/renderWithTheme.native';
+
+describe('<TranslateIcon />', () => {
+  it('should render TranslateIcon', () => {
+    const renderTree = renderWithTheme(
+      <TranslateIcon color="feedback.icon.neutral.intense" size="large" />,
+    ).toJSON();
+    expect(renderTree).toMatchSnapshot();
+  });
+});