+++ packages/blade/src/components/SideNav/types.ts
@@ -4,7 +4,7 @@ import type { StyledPropsBlade } from '~components/Box/styledProps';
 import type { DrawerProps } from '~components/Drawer';
 import type { IconComponent } from '~components/Icons';
 import type { LinkProps } from '~components/Link';
-import type { TooltipProps } from '~components/Tooltip';
+import type { TooltipifyComponentProps } from '~utils/TooltipifyComponent';
 import type { DataAnalyticsAttribute, TestID } from '~utils/types';
 
 type SideNavProps = {
@@ -28,6 +28,15 @@ type SideNavProps = {
   onDismiss?: DrawerProps['onDismiss'];
 
   /**
+   * Callback that gets triggered when L1 is collapsed or expanded.
+   *
+   * This callback gets triggered when you-
+   * - Select the active link changes between L1 and L2 which can collapse or expand the L1
+   * - When you hover / unhover L1 in collapsed state which can temporarily expand the L1
+   */
+  onVisibleLevelChange?: ({ visibleLevel }: { visibleLevel: number }) => void;
+
+  /**
    * Banner slot for usecases like adding Activation Panel
    *
    * **IMPORTANT** Avoid adding promotional items in this
@@ -41,6 +50,13 @@ type SideNavLinkProps = {
    * title of the Link
    */
   title: string;
+
+  /**
+   * description of the Link
+   *
+   * **Note**: Only applicable for L2 items
+   */
+  description?: string;
 
   /**
    * Slot after the title.
@@ -115,7 +131,7 @@ type SideNavLinkProps = {
    * />
    * ```
    */
-  tooltip?: Pick<TooltipProps, 'title' | 'content' | 'onOpenChange'>;
+  tooltip?: TooltipifyComponentProps['tooltip'];
   onClick?: (event: React.MouseEvent) => void;
 } & DataAnalyticsAttribute;
 
@@ -158,6 +174,7 @@ type OnLinkActiveChangeArgs = {
 };
 
 type SideNavContextType = {
+  isL1Hovered?: boolean;
   isL1Collapsed?: boolean;
   setIsL1Collapsed?: (isL1Collapsed: boolean) => void;
   l2PortalContainerRef?: React.RefObject<HTMLDivElement>;